добавил ещё проверок на чат айди

+++ bot.py
@@ -82,31 +82,44 @@ async def command_service(message : types.Message):
     return
 @dp.message_handler(commands=['msg'])
 async def command_msg(message : types.Message):
-    if message.from_user.is_bot != True:
-        msg = message.text.strip('/msg ')
-        await bot_message.send_msg(bot, msg, message.from_id, message)
-        return
+    if message.from_user.id==message.chat.id:
+        if message.from_user.is_bot != True:
+            msg = message.text.strip('/msg ')
+            await bot_message.send_msg(bot, msg, message.from_id, message)
+    else:
+        await message.reply("Для общения с ботом напиши ему в ЛС @%s"%botname)
+        await message.delete()
+    return
 @dp.message_handler(commands=['message'])
 async def command_msg(message : types.Message):
-    if message.from_user.is_bot != True:
-        msg = message.text.strip('/message ')
-        await bot_message.send_msg(bot, msg, message.from_id, message)
-        return
+    if message.from_user.id==message.chat.id:
+        if message.from_user.is_bot != True:
+            msg = message.text.strip('/message ')
+            await bot_message.send_msg(bot, msg, message.from_id, message)
+    else:
+        await message.reply("Для общения с ботом напиши ему в ЛС @%s"%botname)
+        await message.delete()
+    return
 
 
 @dp.message_handler(commands=get_command())
 async def command_client(message : types.Message):
-    comm_data,comm_name = get_command_data(message.text.strip('/'))
-    # print(comm_data)
-    # print(comm_name)
-    if comm_data==0:
-        #start dialog
-        await bot.send_message(message.from_id, dialog.start(comm_name, message.from_id))
-        await message.delete()
+    if message.from_user.id==message.chat.id:
+        if message.from_user.is_bot != True:
+            comm_data,comm_name = get_command_data(message.text.strip('/'))
+            if comm_data==0:
+                #start dialog
+                await bot.send_message(message.from_id, dialog.start(comm_name, message.from_id))
+                await message.delete()
+            else:
+                #start faq
+                await bot.send_message(message.from_id, dialog.faq(comm_name))
+                await message.delete()
     else:
-        #start faq
-        await bot.send_message(message.from_id, dialog.faq(comm_name))
+        await message.reply("Для общения с ботом напиши ему в ЛС @%s"%botname)
         await message.delete()
+    return
+
 
 @dp.message_handler()
 async def echo_send(message : types.Message):
